Update location of python include

+++ tf_sig_build_dockerfiles/setup.python.sh
@@ -31,7 +31,13 @@ popd
 ln -sf /usr/bin/$VERSION /usr/bin/python3
 ln -sf /usr/bin/$VERSION /usr/bin/python
 ln -sf /usr/lib/$VERSION /usr/lib/tf_python
-ln -sf /usr/include/$VERSION /usr/local/include/$VERSION
+
+# Python 3.10 include headers fix:
+# sysconfig.get_path('include') incorrectly points to /usr/local/include/python
+# map /usr/include/python3.10 to /usr/local/include/python3.10
+if [[ -f "/usr/local/include/$VERSION" ]]; then
+  ln -sf /usr/include/$VERSION /usr/local/include/$VERSION
+fi
 
 # Install pip
 curl https://bootstrap.pypa.io/get-pip.py -o get-pip.py
